requirements.txt
numpy
pandas
python-dotenv
llama-index
llama-hub
setuptools
pypdf
